requirements.txt
Django==4.1.5
django-crum==0.7.9
flake8==5.0.4
